Add type annotations

+++ modules/db_utils.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from typing import Callable, List, Optional, Set, Tuple
-import apsw
+import apsw  # type: ignore
 from apsw import Error
 import logging
 from hashlib import sha256
@@ -47,39 +47,41 @@ def create_connection(filename: str) -> Optional[apsw.Connection]:
 
 def create_ips_table(filename: str) -> None:
     conn = create_connection(filename)
-    try:
-        with conn:
-            cur = conn.cursor()
-            cur.execute(
-                f"""CREATE TABLE IF NOT EXISTS urls (
-                           url text,
-                           lastGoogleMalicious integer,
-                           lastYandexMalicious integer,
-                           hash blob
-                           )"""
-            )
-    except Error as e:
-        logging.error(f"filename:{filename} {e}")
-    conn.close()
+    if conn is not None:
+        try:
+            with conn:
+                cur = conn.cursor()
+                cur.execute(
+                    f"""CREATE TABLE IF NOT EXISTS urls (
+                            url text,
+                            lastGoogleMalicious integer,
+                            lastYandexMalicious integer,
+                            hash blob
+                            )"""
+                )
+        except Error as e:
+            logging.error(f"filename:{filename} {e}")
+        conn.close()
 
 
 def create_urls_table(filename: str) -> None:
     conn = create_connection(filename)
-    try:
-        with conn:
-            cur = conn.cursor()
-            cur.execute(
-                f"""CREATE TABLE IF NOT EXISTS urls (
-                           url text UNIQUE,
-                           lastListed integer,
-                           lastGoogleMalicious integer,
-                           lastYandexMalicious integer,
-                           hash blob
-                           )"""
-            )
-    except Error as e:
-        logging.error(f"filename:{filename} {e}")
-    conn.close()
+    if conn is not None:
+        try:
+            with conn:
+                cur = conn.cursor()
+                cur.execute(
+                    f"""CREATE TABLE IF NOT EXISTS urls (
+                            url text UNIQUE,
+                            lastListed integer,
+                            lastGoogleMalicious integer,
+                            lastYandexMalicious integer,
+                            hash blob
+                            )"""
+                )
+        except Error as e:
+            logging.error(f"filename:{filename} {e}")
+        conn.close()
 
 
 def compute_url_hash(url: str) -> bytes:
@@ -99,42 +101,43 @@ def add_IPs(filename: str, first_octet: int) -> None:
     1 file for each bit in first octet.
     """
     conn = create_connection(filename)
-    ips_to_generate = 2 ** 24
-    try:
-        # Check if there are 2 ** 24 ips in DB
-        with conn:
-            cur = conn.cursor()
-            cur.execute("SELECT MAX(_rowid_) FROM urls")
-            number_of_ipv4_addresses = cur.fetchall()[0][0]
-        if number_of_ipv4_addresses != ips_to_generate:
-            # If DB does not have 2 ** 24 IPs, delete all rows from ipv4_ urls table and regenerate IPs
-            logging.info(
-                f"INSERT {ips_to_generate} ipv4 addresses to urls table of {filename}..."
-            )
+    if conn is not None:
+        ips_to_generate = 2 ** 24
+        try:
+            # Check if there are 2 ** 24 ips in DB
             with conn:
                 cur = conn.cursor()
-                cur.execute("DELETE FROM urls")
-            with conn:
-                cur = conn.cursor()
-                for int_addr in range(ips_to_generate):
-                    cur.execute(
-                        f"""
-                    INSERT INTO urls (url,hash)
-                    VALUES (?,?)
-                    """,
-                        int_addr_to_ip_and_hash(int_addr + (2 ** 24) * first_octet),
-                    )
-
+                cur.execute("SELECT MAX(_rowid_) FROM urls")
+                number_of_ipv4_addresses = cur.fetchall()[0][0]
+            if number_of_ipv4_addresses != ips_to_generate:
+                # If DB does not have 2 ** 24 IPs, delete all rows from ipv4_ urls table and regenerate IPs
                 logging.info(
-                    f"INSERT {ips_to_generate} ipv4 addresses to urls table of {filename}...[DONE]"
+                    f"INSERT {ips_to_generate} ipv4 addresses to urls table of {filename}..."
                 )
-    except Error as e:
-        logging.error(f"filename:{filename} {e}")
-    conn.close()
+                with conn:
+                    cur = conn.cursor()
+                    cur.execute("DELETE FROM urls")
+                with conn:
+                    cur = conn.cursor()
+                    for int_addr in range(ips_to_generate):
+                        cur.execute(
+                            f"""
+                        INSERT INTO urls (url,hash)
+                        VALUES (?,?)
+                        """,
+                            int_addr_to_ip_and_hash(int_addr + (2 ** 24) * first_octet),
+                        )
+
+                    logging.info(
+                        f"INSERT {ips_to_generate} ipv4 addresses to urls table of {filename}...[DONE]"
+                    )
+        except Error as e:
+            logging.error(f"filename:{filename} {e}")
+        conn.close()
 
 
 def add_URLs(
-    url_list_fetcher: Callable[[Optional[str]], List[str]],
+    url_list_fetcher: Callable[..., List[str]],
     updateTime: int,
     filename: str,
     filepath: Optional[str] = None,
@@ -146,30 +149,31 @@ def add_URLs(
     urls = url_list_fetcher() if filepath == None else url_list_fetcher(filepath)
     lastListed = updateTime
     conn = create_connection(filename)
-    try:
-        with conn:
-            cur = conn.cursor()
-            logging.info(
-                f"Performing INSERT-UPDATE URLs to urls table of {filename}..."
-            )
-
-            for url_batch in urls:
-                cur.executemany(
-                    f"""
-                INSERT INTO urls (url, lastListed, hash)
-                VALUES (?,?,?)
-                ON CONFLICT(url)
-                DO UPDATE SET lastListed=excluded.lastListed
-                """,
-                    ((url, lastListed, compute_url_hash(url)) for url in url_batch),
+    if conn is not None:
+        try:
+            with conn:
+                cur = conn.cursor()
+                logging.info(
+                    f"Performing INSERT-UPDATE URLs to urls table of {filename}..."
                 )
 
-            logging.info(
-                f"Performing INSERT-UPDATE URLs to urls table of {filename}...[DONE]"
-            )
-    except Error as e:
-        logging.error(f"filename:{filename} {e}")
-    conn.close()
+                for url_batch in urls:
+                    cur.executemany(
+                        f"""
+                    INSERT INTO urls (url, lastListed, hash)
+                    VALUES (?,?,?)
+                    ON CONFLICT(url)
+                    DO UPDATE SET lastListed=excluded.lastListed
+                    """,
+                        ((url, lastListed, compute_url_hash(url)) for url in url_batch),
+                    )
+
+                logging.info(
+                    f"Performing INSERT-UPDATE URLs to urls table of {filename}...[DONE]"
+                )
+        except Error as e:
+            logging.error(f"filename:{filename} {e}")
+        conn.close()
 
 
 def add_maliciousHashPrefixes(hash_prefixes: Set[bytes], vendor: str) -> None:
@@ -178,20 +182,26 @@ def add_maliciousHashPrefixes(hash_prefixes: Set[bytes], vendor: str) -> None:
     """
     logging.info(f"Updating DB with {vendor} malicious URL hashes")
     conn = create_connection("malicious")
-    try:
-        with conn:
-            cur = conn.cursor()
-            cur.execute("DELETE FROM maliciousHashPrefixes WHERE vendor = ?", (vendor,))
-            cur.executemany(
-                """
-                INSERT INTO maliciousHashPrefixes (hashPrefix,prefixSize,vendor)
-                VALUES (?, ?, ?)
-                """,
-                ((hashPrefix, len(hashPrefix), vendor) for hashPrefix in hash_prefixes),
-            )
-    except Error as e:
-        logging.error(f"vendor:{vendor} {e}")
-    conn.close()
+    if conn is not None:
+        try:
+            with conn:
+                cur = conn.cursor()
+                cur.execute(
+                    "DELETE FROM maliciousHashPrefixes WHERE vendor = ?", (vendor,)
+                )
+                cur.executemany(
+                    """
+                    INSERT INTO maliciousHashPrefixes (hashPrefix,prefixSize,vendor)
+                    VALUES (?, ?, ?)
+                    """,
+                    (
+                        (hashPrefix, len(hashPrefix), vendor)
+                        for hashPrefix in hash_prefixes
+                    ),
+                )
+        except Error as e:
+            logging.error(f"vendor:{vendor} {e}")
+        conn.close()
 
 
 def get_matching_hashPrefix_urls(
@@ -199,46 +209,49 @@ def get_matching_hashPrefix_urls(
 ) -> List[str]:
     conn = create_connection(filename)
     urls = []
-    try:
-        with conn:
-            cur = conn.cursor()
-            cur = cur.execute(
-                f"ATTACH database 'databases{os.sep}malicious.db' as malicious"
+    if conn is not None:
+        try:
+            with conn:
+                cur = conn.cursor()
+                cur = cur.execute(
+                    f"ATTACH database 'databases{os.sep}malicious.db' as malicious"
+                )
+                cur = cur.execute(
+                    f"""SELECT url FROM urls 
+                    WHERE substring(urls.hash,1,?) IN (SELECT hashPrefix FROM malicious.maliciousHashPrefixes
+                    WHERE vendor = ?)""",
+                    (prefixSize, vendor),
+                )
+                urls = [x[0] for x in cur.fetchall()]
+            with conn:
+                cur = conn.cursor()
+                cur = cur.execute("DETACH database malicious")
+        except Error as e:
+            logging.error(
+                f"filename:{filename} prefixSize:{prefixSize} vendor:{vendor} {e}"
             )
-            cur = cur.execute(
-                f"""SELECT url FROM urls 
-                WHERE substring(urls.hash,1,?) IN (SELECT hashPrefix FROM malicious.maliciousHashPrefixes
-                WHERE vendor = ?)""",
-                (prefixSize, vendor),
-            )
-            urls = [x[0] for x in cur.fetchall()]
-        with conn:
-            cur = conn.cursor()
-            cur = cur.execute("DETACH database malicious")
-    except Error as e:
-        logging.error(
-            f"filename:{filename} prefixSize:{prefixSize} vendor:{vendor} {e}"
-        )
-    conn.close()
+        conn.close()
 
     return urls
 
 
 def retrieve_vendor_prefixSizes(vendor: str) -> List[int]:
-    conn = create_connection("malicious")
     prefixSizes = []
-    try:
-        with conn:
-            # Find all prefixSizes
-            cur = conn.cursor()
-            cur = cur.execute(
-                "SELECT DISTINCT prefixSize FROM maliciousHashPrefixes WHERE vendor = ?",
-                (vendor,),
-            )
-            prefixSizes = [x[0] for x in cur.fetchall()]
-    except Error as e:
-        logging.error(f"vendor: {vendor} {e}")
-    conn.close()
+
+    conn = create_connection("malicious")
+    if conn is not None:
+        try:
+            with conn:
+                # Find all prefixSizes
+                cur = conn.cursor()
+                cur = cur.execute(
+                    "SELECT DISTINCT prefixSize FROM maliciousHashPrefixes WHERE vendor = ?",
+                    (vendor,),
+                )
+                prefixSizes = [x[0] for x in cur.fetchall()]
+        except Error as e:
+            logging.error(f"vendor: {vendor} {e}")
+        conn.close()
     return prefixSizes
 
 
@@ -249,19 +262,20 @@ def create_maliciousHashPrefixes_table() -> None:
     :return:
     """
     conn = create_connection("malicious")
-    try:
-        with conn:
-            cur = conn.cursor()
-            cur.execute(
-                """CREATE TABLE IF NOT EXISTS maliciousHashPrefixes (
-                                            hashPrefix blob,
-                                            prefixSize integer,
-                                            vendor text
-                                            )"""
-            )
-    except Error as e:
-        logging.error(f"{e}")
-    conn.close()
+    if conn is not None:
+        try:
+            with conn:
+                cur = conn.cursor()
+                cur.execute(
+                    """CREATE TABLE IF NOT EXISTS maliciousHashPrefixes (
+                                                hashPrefix blob,
+                                                prefixSize integer,
+                                                vendor text
+                                                )"""
+                )
+        except Error as e:
+            logging.error(f"{e}")
+        conn.close()
 
 
 def initialise_database(filenames: List[str], mode: str) -> None:
@@ -299,27 +313,28 @@ def update_malicious_URLs(
     if vendor not in vendorToColumn:
         raise ValueError('vendor must be "Google" or "Yandex"')
     conn = create_connection(filename)
-    try:
-        batch_size = 30_000
-        malicious_url_batches = chunked(malicious_urls, batch_size)
-        for malicious_url_batch in malicious_url_batches:
-            malicious_url_batch_length = len(malicious_url_batch)
-            with conn:
-                cur = conn.cursor()
-                cur.execute(
-                    f"""
-                    UPDATE urls
-                    SET {vendorToColumn[vendor]} = ?
-                    WHERE url IN ({','.join('?'*malicious_url_batch_length)})
-                    """,
-                    (updateTime, *malicious_url_batch),
-                )
-        logging.info(
-            f"Updating {filename} DB with verified {vendor} malicious URLs...[DONE]"
-        )
-    except Error as e:
-        logging.error(f"vendor:{vendor} filename:{filename} {e}")
-    conn.close()
+    if conn is not None:
+        try:
+            batch_size = 30_000
+            malicious_url_batches = chunked(malicious_urls, batch_size)
+            for malicious_url_batch in malicious_url_batches:
+                malicious_url_batch_length = len(malicious_url_batch)
+                with conn:
+                    cur = conn.cursor()
+                    cur.execute(
+                        f"""
+                        UPDATE urls
+                        SET {vendorToColumn[vendor]} = ?
+                        WHERE url IN ({','.join('?'*malicious_url_batch_length)})
+                        """,
+                        (updateTime, *malicious_url_batch),
+                    )
+            logging.info(
+                f"Updating {filename} DB with verified {vendor} malicious URLs...[DONE]"
+            )
+        except Error as e:
+            logging.error(f"vendor:{vendor} filename:{filename} {e}")
+        conn.close()
 
 
 def retrieve_malicious_URLs(urls_filenames: List[str]) -> List[str]:
@@ -328,28 +343,29 @@ def retrieve_malicious_URLs(urls_filenames: List[str]) -> List[str]:
     """
 
     def retrieve_malicious_URLs_(filename: str) -> Set[str]:
-        malicious_urls = set()
+        malicious_urls: Set[str] = set()
         conn = create_connection(filename)
-        try:
-            with conn:
-                cur = conn.cursor()
-                # Most recent lastGoogleMalicious timestamp
-                cur.execute(f"SELECT MAX(lastGoogleMalicious) FROM urls")
-                lastGoogleMalicious = [x[0] for x in cur.fetchall()][0]
-                # Most recent lastYandexMalicious timestamp
-                cur.execute(f"SELECT MAX(lastYandexMalicious) FROM urls")
-                lastYandexMalicious = [x[0] for x in cur.fetchall()][0]
-                cur.execute(
-                    f"""
-        SELECT url FROM urls
-        WHERE lastGoogleMalicious = ? OR lastYandexMalicious = ?
-        """,
-                    (lastGoogleMalicious, lastYandexMalicious),
-                )
-                malicious_urls.update((x[0] for x in cur.fetchall()))
-        except Error as e:
-            logging.error(f"filename:{filename} {e}")
-        conn.close()
+        if conn is not None:
+            try:
+                with conn:
+                    cur = conn.cursor()
+                    # Most recent lastGoogleMalicious timestamp
+                    cur.execute(f"SELECT MAX(lastGoogleMalicious) FROM urls")
+                    lastGoogleMalicious = [x[0] for x in cur.fetchall()][0]
+                    # Most recent lastYandexMalicious timestamp
+                    cur.execute(f"SELECT MAX(lastYandexMalicious) FROM urls")
+                    lastYandexMalicious = [x[0] for x in cur.fetchall()][0]
+                    cur.execute(
+                        f"""
+            SELECT url FROM urls
+            WHERE lastGoogleMalicious = ? OR lastYandexMalicious = ?
+            """,
+                        (lastGoogleMalicious, lastYandexMalicious),
+                    )
+                    malicious_urls.update((x[0] for x in cur.fetchall()))
+            except Error as e:
+                logging.error(f"filename:{filename} {e}")
+            conn.close()
 
         return malicious_urls
